requirements.txt
Flask==2.2.5
requests==2.31.0
gunicorn==21.2.0
google-cloud-aiplatform
google-cloud-secret-manager